+++ release-notes.md
@@ -1,6 +1,8 @@
 Version 1.4.1 (unreleased)
 --------------
 * Added some copy constructors to Vec3d, Vec4d, and Quatd.
+* Added getters for the start/end pos, rotation, visibility of 
+    PositionTransition3d and PositionTransition3f
 
 
 Version 1.4.0 (latest)